This is a Revert commit. It will patch the problem that your commits made.
Please, commit again your code in the SRoad, without cause conflicts.

Notes: To bypass the conflicts and every core error with constructors
and method's arguments, use the Polymorphism idea of JAVA. In other
words, write again a 2nd constructor or method.
In this case we'll have to time to merge the functionality of the new
methods/constructors without compiling errors.