Framework completely changed - Read the description
Completely deleted classes:
- Node.java
- NodeManager.java
- TestingNodeNetwork.java

Edited:
~ All the rest, except of the TraffictLights and TrafficLightsGUI
(Which both of them should be merged!!)

Added:
+ TrafficManagement.java

A quick review of the code:
Due to the fact that every node has ONLY one terrain there is not a
reason to have an extra “Node” object in the memory.

Instead of that, think that every Road (Interface Terrain) is the new
Node. This thought will make the rest much more simple to understand.

In the new files you find the following idea.

Each Node  =  a Road (Terrain)
In each Road there are 2 ArrayList<objects>  which are the  new node
stacks.
There are 2 stacks because we have 2 different flows in each Road.
One flow has direction —> and the other has direction <—

Also, there are some changes and pseudo code is comments in some parts
about the policies and the car flow.

Important:
1 road will have ONLY one Traffic Light by each direction, which is
located on top of the local stack.
Code:  backwardList.get(0) = Traffic Light of the <— direction of the
road.

Important 2: Don’t add in Junctions any Traffic light. and of course in
any turn (Curve road). There is no point to do there.

Additionally,

Idea:
2 options for the Thread implementation:

1) Each Car flow is a thread, but in the car flow is located a local
variable for the current terrain. Previous each car flow had a variable
for the Vehicles.
So, each car.move(arg1,arg2) will be done periodically by car.
I added some first version code.

2) Each car flow again is a thread with a local variable but each car
flow will create a thread of cars and the implementation will be more
or less like the moving that we have already online. However, here we
have
( Thread.CarFlows x Thread.Vehicles ) .

About the threads:
Logically, both of the 2 stacks of each road should be synchronised,
because each thread will write to the stack free.

For the option 1 we need Thread pool in the TrafficManagement. Thread
pool size will be equal to  the aTerrainList.size().

For the option 2, thread pool in the Traffic Management, and maybe each
Car Flow should save the threads somewhere.

Of course there is no moving there, because the Vehicle.move() is never
called. This move() should be changed.

Don’t Forget:
Download the /src/ locally and run it.
Review the code, and the comment and feel free to make questions
improvements.